+++ app/controllers/rails/monitoring/application_controller.rb
@@ -1,0 +1,7 @@
+module Rails
+  module Monitoring
+    class ApplicationController < Rails::Monitoring.parent_controller.constantize
+      protect_from_forgery with: :exception
+    end
+  end
+end
