--- app/controllers/sidekiq_monitoring/application_controller.rb
@@ -1,5 +1,0 @@
-module SidekiqMonitoring
-  class ApplicationController < SidekiqMonitoring.parent_controller.constantize
-    protect_from_forgery with: :exception
-  end
-end
