+++ app/jobs/rails/monitoring/application_job.rb
@@ -1,0 +1,6 @@
+module Rails
+  module Monitoring
+    class ApplicationJob < ActiveJob::Base
+    end
+  end
+end
