--- app/jobs/sidekiq_monitoring/application_job.rb
@@ -1,4 +1,0 @@
-module SidekiqMonitoring
-  class ApplicationJob < ActiveJob::Base
-  end
-end
