--- app/jobs/sidekiq_monitoring/refresh_status_job.rb
@@ -1,9 +1,0 @@
-module SidekiqMonitoring
-  class RefreshStatusJob < ApplicationJob
-    def perform
-      Sidekiq.redis do |conn|
-        conn.set('monitoring:timestamp:sidekiq_performed', Time.current.to_s(:db))
-      end
-    end
-  end
-end
