+++ bin/rails
@@ -3,7 +3,7 @@
 # installed from the root of your application.
 
 ENGINE_ROOT = File.expand_path('../..', __FILE__)
-ENGINE_PATH = File.expand_path('../../lib/sidekiq_monitoring/engine', __FILE__)
+ENGINE_PATH = File.expand_path('../../lib/rails/monitoring/engine', __FILE__)
 APP_PATH = File.expand_path('../../test/dummy/config/application', __FILE__)
 
 # Set up gems listed in the Gemfile.
