+++ config/routes.rb
@@ -1,4 +1,4 @@
-SidekiqMonitoring::Engine.routes.draw do
+Rails::Monitoring::Engine.routes.draw do
   controller :status do
     get :status
   end
