+++ lib/rails/monitoring/engine.rb
@@ -1,0 +1,10 @@
+require 'sidekiq'
+require 'sidekiq/api'
+
+module Rails
+  module Monitoring
+    class Engine < ::Rails::Engine
+      isolate_namespace Rails::Monitoring
+    end
+  end
+end
