+++ lib/rails/monitoring/version.rb
@@ -1,0 +1,5 @@
+module Rails
+  module Monitoring
+    VERSION = '0.2.0'.freeze
+  end
+end
