--- lib/sidekiq_monitoring_bitcrowd/engine.rb
@@ -1,8 +1,0 @@
-require 'sidekiq'
-require 'sidekiq/api'
-
-module SidekiqMonitoring
-  class Engine < ::Rails::Engine
-    isolate_namespace SidekiqMonitoring
-  end
-end
