--- lib/sidekiq_monitoring_bitcrowd/version.rb
@@ -1,3 +1,0 @@
-module SidekiqMonitoring
-  VERSION = '0.2.0'.freeze
-end
