--- test/dummy/app/assets/config/manifest.js
@@ -1,5 +1,0 @@
-
-//= link_tree ../images
-//= link_directory ../javascripts .js
-//= link_directory ../stylesheets .css
-//= link sidekiq_monitoring_manifest.js
