+++ test/dummy/config/initializers/monitoring.rb
@@ -1,0 +1,3 @@
+Rails::Monitoring.parent_controller = 'SomeController'
+Rails::Monitoring.http_auth_name = 'some_user'
+Rails::Monitoring.http_auth_password = 'some_password'
