--- test/dummy/config/initializers/sidekiq_monitoring.rb
@@ -1,3 +1,0 @@
-SidekiqMonitoring.parent_controller = 'SomeController'
-SidekiqMonitoring.http_auth_name = 'some_user'
-SidekiqMonitoring.http_auth_password = 'some_password'
